Move ConfigValueBean into it's own class
